refactoring python module and silencing unneeded output when using modules pythonically
