+++ dnsrecon/__init__.py
@@ -5,7 +5,9 @@ dnsrecon python module adaptation
 '''
 
 from .lib.crtenum import scrape_crtsh
-from .lib.yandexenum import *
+from .lib.bingenum import scrape_bing
+from .lib.yandexenum import scrape_yandex
 from .dnsrecon import *
 from .msf_print import *
 from .whois import *
+from .lib.output import *
